+++ common/src/main/scala/com/couchmate/common/models/SportOrganization.scala
@@ -5,7 +5,7 @@ import play.api.libs.json.{Json, OFormat}
 case class SportOrganization(
   sportOrganizationId: Option[Long],
   extSportId: Long,
-  extOrgId: Option[Int],
+  extOrgId: Option[Long],
   sportName: String,
   orgName: Option[String],
 ) extends Product with Serializable
